dash: split out TimeInput into @shared/components
