+++ shared/components/package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@headlessui/react": "1.7.11",
+    "@heroicons/react": "2.0.17",
     "@shared/datetime": "workspace:*",
     "classnames": "^2.3.2",
     "react": "18.2.0",
